requirements/local.txt
anyjson==0.3.3
boto3==1.6.21
celery==4.1.0
circus==0.14.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-ckeditor==5.4.0
django-cors-headers==2.2.0
#--- ONLY FOR LOCAL
django-debug-panel==0.8.3
#--- ONLY FOR LOCAL
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-extensions==2.0.6
django-smtp-ssl==1.0
django-user-agents==0.3.2
djangorestframework==3.9.2
django-rest-swagger==2.1.2
#--- ONLY FOR LOCAL
Fabric3==1.14.post1
#--- ONLY FOR LOCAL
flake8==3.5.0
#--- ONLY FOR LOCAL
flake8-colors==0.1.6
gunicorn==19.7.1
#--- ONLY FOR LOCAL
ipdb==0.11
kv-settings==2.0.0
Pillow==5.0.0
pytz==2018.3
redis==2.10.6
requests==2.21.0
# simple_django_logger==0.1.3
#--- ONLY FOR LOCAL
Sphinx==1.7.2
user-agents==1.1.0
